Address agent workflow review feedback
