--- packages/agent-sdk-test/tests/agent-sdk.test.ts
@@ -1,470 +1,0 @@
-import { describe, expect, it } from "vitest"
-import { createScope, flow, tag, tags, typed } from "@pumped-fn/lite"
-import {
-  CliWorkerError,
-  SuspendSignal,
-  agent as agentRuntime,
-  claudeCliWorker,
-  cliWorker,
-  codexCliWorker,
-  derivedMaterial,
-  extension as agentExtension,
-  material,
-  patchMaterial,
-  run,
-  step,
-  workers,
-  workflow as workflowRuntime,
-  workflowExtension,
-  workerRegistry,
-} from "@pumped-fn/agent-sdk"
-import { replay, run as replayRun, suspend } from "@pumped-fn/lite-extension-suspense"
-import {
-  InMemoryAgentEventLog,
-  InMemorySuspenseEventLog,
-  agent,
-  suspense,
-} from "../src/index"
-
-describe("agent sdk", () => {
-  it("replays standalone suspense steps without agent config", async () => {
-    const log = new InMemorySuspenseEventLog()
-    const { extension } = suspense({ log })
-    const scope = createScope({ extensions: [extension] })
-    await scope.ready
-    let calls = 0
-    const step = flow({
-      name: "standalone-step",
-      parse: typed<number>(),
-      tags: [replay(true)],
-      factory: (ctx) => {
-        calls++
-        return ctx.input + 1
-      },
-    })
-
-    const ctx1 = scope.createContext(replayRun({ taskId: "sync-a", runId: "run-a" }))
-    expect(await ctx1.exec({ flow: step, input: 1 })).toBe(2)
-    await ctx1.close()
-
-    const ctx2 = scope.createContext(replayRun({ taskId: "sync-a", runId: "run-a" }))
-    expect(await ctx2.exec({ flow: step, input: 100 })).toBe(2)
-    await ctx2.close()
-    expect(calls).toBe(1)
-  })
-
-  it("suspends standalone suspense steps and resumes from resolved value", async () => {
-    const log = new InMemorySuspenseEventLog()
-    const { extension } = suspense({ log })
-    const scope = createScope({ extensions: [extension] })
-    await scope.ready
-    const externalSync = flow({
-      name: "external-sync",
-      tags: [suspend(true)],
-      factory: () => "unreachable",
-    })
-
-    const ctx1 = scope.createContext(replayRun({ taskId: "sync-b", runId: "run-b" }))
-    await expect(ctx1.exec({ flow: externalSync })).rejects.toBeInstanceOf(SuspendSignal)
-    await ctx1.close({ ok: false, error: new Error("suspended") })
-
-    const pending = log.entries().find((entry) => entry.status === "pending")
-    expect(pending?.targetName).toBe("external-sync")
-    if (!pending) throw new Error("external-sync step did not suspend")
-    await log.resolve(pending.key, "synced")
-
-    const ctx2 = scope.createContext(replayRun({ taskId: "sync-b", runId: "run-b" }))
-    expect(await ctx2.exec({ flow: externalSync })).toBe("synced")
-    await ctx2.close()
-  })
-
-  it("memoizes workflow runtime tag with workflow extension", async () => {
-    const log = new InMemoryAgentEventLog()
-    const extension = workflowExtension({ log })
-    const scope = createScope({ extensions: [extension] })
-    await scope.ready
-    let calls = 0
-    const root = flow({
-      name: "workflow-root",
-      parse: typed<number>(),
-      tags: [step({ workflow: true })],
-      deps: { workflow: tags.required(workflowRuntime) },
-      factory: (ctx, { workflow }) => {
-        calls++
-        return `${workflow.taskId}:${workflow.runId}:${ctx.input}`
-      },
-    })
-
-    const ctx1 = scope.createContext({ tags: [run({ taskId: "workflow-task", runId: "workflow-run" })] })
-    expect(await ctx1.exec({ flow: root, input: 1 })).toBe("workflow-task:workflow-run:1")
-    await ctx1.close()
-
-    const ctx2 = scope.createContext({ tags: [run({ taskId: "workflow-task", runId: "workflow-run" })] })
-    expect(await ctx2.exec({ flow: root, input: 2 })).toBe("workflow-task:workflow-run:1")
-    await ctx2.close()
-    expect(calls).toBe(1)
-  })
-
-  it("rejects workflow dependency without workflow extension", async () => {
-    const root = flow({
-      name: "unguarded-workflow",
-      tags: [step({ workflow: true })],
-      deps: { workflow: tags.required(workflowRuntime) },
-      factory: (_ctx, { workflow }) => workflow.taskId,
-    })
-    const scope = createScope()
-    const ctx = scope.createContext({ tags: [run({ taskId: "workflow-missing-extension", runId: "run" })] })
-    await expect(ctx.exec({ flow: root })).rejects.toThrow('Tag "workflow.runtime" not found')
-    await ctx.close({ ok: false, error: new Error("expected") })
-    await scope.dispose()
-  })
-
-  it("provides workflow runtime deps without workflow memoization", async () => {
-    const log = new InMemoryAgentEventLog()
-    const extension = workflowExtension({ log })
-    const scope = createScope({ extensions: [extension] })
-    await scope.ready
-    let calls = 0
-    const root = flow({
-      name: "workflow-runtime-only",
-      parse: typed<number>(),
-      deps: { workflow: tags.required(workflowRuntime) },
-      factory: (ctx, { workflow }) => {
-        calls++
-        return `${workflow.taskId}:${ctx.input}`
-      },
-    })
-
-    const ctx1 = scope.createContext({ tags: [run({ taskId: "runtime-task", runId: "runtime-run" })] })
-    expect(await ctx1.exec({ flow: root, input: 1 })).toBe("runtime-task:1")
-    await ctx1.close()
-
-    const ctx2 = scope.createContext({ tags: [run({ taskId: "runtime-task", runId: "runtime-run" })] })
-    expect(await ctx2.exec({ flow: root, input: 2 })).toBe("runtime-task:2")
-    await ctx2.close()
-    expect(calls).toBe(2)
-  })
-
-  it("memoizes completed workflow execution", async () => {
-    const { extensions } = agent()
-    const scope = createScope({ extensions })
-    await scope.ready
-    let calls = 0
-    const worker = flow({
-      name: "memo-worker",
-      parse: typed<number>(),
-      tags: [step({ workflow: true })],
-      factory: (ctx) => {
-        calls++
-        return ctx.input * 2
-      },
-    })
-    const root = flow({
-      name: "memo-root",
-      parse: typed<number>(),
-      tags: [step({ workflow: true })],
-      deps: { workflow: tags.required(workflowRuntime) },
-      factory: async (ctx, { workflow }) => {
-        expect(workflow.taskId).toBe("task-a")
-        return ctx.exec({ flow: worker, input: ctx.input })
-      },
-    })
-
-    const ctx1 = scope.createContext({ tags: [run({ taskId: "task-a", runId: "run-a" })] })
-    expect(await ctx1.exec({ flow: root, input: 3 })).toBe(6)
-    await ctx1.close()
-
-    const ctx2 = scope.createContext({ tags: [run({ taskId: "task-a", runId: "run-a" })] })
-    expect(await ctx2.exec({ flow: root, input: 3 })).toBe(6)
-    await ctx2.close()
-
-    expect(calls).toBe(1)
-  })
-
-  it("replays completed flow before resolving deps", async () => {
-    const gate = tag<string>({ label: "agent.replay.gate" })
-    const { extensions } = agent()
-    const scope = createScope({ extensions })
-    await scope.ready
-    let calls = 0
-    const worker = flow({
-      name: "replay-before-deps",
-      tags: [step({ workflow: true })],
-      deps: { gate: tags.required(gate) },
-      factory: (_ctx, { gate }) => {
-        calls++
-        return `ok:${gate}`
-      },
-    })
-
-    const ctx1 = scope.createContext({
-      tags: [run({ taskId: "task-replay-deps", runId: "run-replay-deps" }), gate("first")],
-    })
-    expect(await ctx1.exec({ flow: worker })).toBe("ok:first")
-    await ctx1.close()
-
-    const ctx2 = scope.createContext({ tags: [run({ taskId: "task-replay-deps", runId: "run-replay-deps" })] })
-    expect(await ctx2.exec({ flow: worker })).toBe("ok:first")
-    await ctx2.close()
-    expect(calls).toBe(1)
-  })
-
-  it("suspends durable work and replays memoized steps on resolution", async () => {
-    const log = new InMemoryAgentEventLog()
-    const { extensions } = agent({ log })
-    const scope = createScope({ extensions })
-    await scope.ready
-    let expensiveCalls = 0
-    const expensive = flow({
-      name: "expensive",
-      tags: [step({ workflow: true })],
-      factory: () => {
-        expensiveCalls++
-        return "ready"
-      },
-    })
-    const approve = flow({
-      name: "approve",
-      tags: [step({ durable: true })],
-      factory: () => "unreachable",
-    })
-    const root = flow({
-      name: "approval-root",
-      tags: [step({ workflow: true })],
-      factory: async (ctx) => {
-        const first = await ctx.exec({ flow: expensive })
-        const decision = await ctx.exec({ flow: approve })
-        return `${first}:${decision}`
-      },
-    })
-
-    const ctx1 = scope.createContext({ tags: [run({ taskId: "task-b", runId: "run-b" })] })
-    await expect(ctx1.exec({ flow: root })).rejects.toBeInstanceOf(SuspendSignal)
-    await ctx1.close({ ok: false, error: new Error("suspended") })
-
-    const pending = log.entries().find((entry) => entry.status === "pending")
-    expect(pending?.targetName).toBe("approve")
-    if (!pending) throw new Error("approve step did not suspend")
-    await log.resolve(pending.key, "yes")
-
-    const ctx2 = scope.createContext({ tags: [run({ taskId: "task-b", runId: "run-b" })] })
-    expect(await ctx2.exec({ flow: root })).toBe("ready:yes")
-    await ctx2.close()
-    expect(expensiveCalls).toBe(1)
-  })
-
-  it("suspends durable work before resolving deps", async () => {
-    const gate = tag<string>({ label: "agent.durable.gate" })
-    const { extensions } = agent()
-    const scope = createScope({ extensions })
-    await scope.ready
-    const approve = flow({
-      name: "durable-before-deps",
-      deps: { gate: tags.required(gate) },
-      factory: (_ctx, { gate }) => gate,
-    })
-
-    const ctx = scope.createContext({ tags: [run({ taskId: "task-durable-deps", runId: "run-durable-deps" })] })
-    await expect(ctx.exec({ flow: approve, tags: [step({ durable: true })] })).rejects.toBeInstanceOf(SuspendSignal)
-    await ctx.close({ ok: false, error: new Error("suspended") })
-  })
-
-  it("delegates by worker registry and routes remote workers", async () => {
-    const { extensions } = agent()
-    const scope = createScope({ extensions })
-    await scope.ready
-    const worker = flow({
-      name: "upper",
-      parse: typed<{ text: string }>(),
-      tags: [step({ remote: true, kind: "code" })],
-      factory: (ctx) => ctx.input.text.toUpperCase(),
-    })
-    const registry = workerRegistry([worker])
-    const root = flow({
-      name: "delegate-root",
-      parse: typed<{ text: string }>(),
-      tags: [step({ workflow: true }), workers(registry)],
-      deps: { agent: tags.required(agentRuntime) },
-      factory: (ctx, { agent }) => agent.delegate<string, { text: string }>("upper", ctx.input),
-    })
-    const ctx = scope.createContext({ tags: [run({ taskId: "task-c", runId: "run-c" })] })
-    expect(await ctx.exec({ flow: root, input: { text: "works" } })).toBe("WORKS")
-    await ctx.close()
-  })
-
-  it("delegates by worker registry tag", async () => {
-    const { extensions } = agent()
-    const scope = createScope({ extensions })
-    await scope.ready
-    const worker = flow({
-      name: "reverse",
-      parse: typed<{ text: string }>(),
-      factory: (ctx) => ctx.input.text.split("").reverse().join(""),
-    })
-    const registry = workerRegistry([worker])
-    const root = flow({
-      name: "delegate-root-tag",
-      parse: typed<{ text: string }>(),
-      tags: [step({ workflow: true })],
-      deps: { agent: tags.required(agentRuntime) },
-      factory: (ctx, { agent }) => agent.delegate<string, { text: string }>("reverse", ctx.input),
-    })
-    const ctx = scope.createContext({
-      tags: [run({ taskId: "task-c-tag", runId: "run-c-tag" }), workers(registry)],
-    })
-    expect(await ctx.exec({ flow: root, input: { text: "works" } })).toBe("skrow")
-    await ctx.close()
-  })
-
-  it("rejects agent dependency without agent extension", async () => {
-    const worker = flow({
-      name: "unguarded-worker",
-      parse: typed<{ text: string }>(),
-      factory: (ctx) => ctx.input.text,
-    })
-    const root = flow({
-      name: "unguarded-root",
-      parse: typed<{ text: string }>(),
-      tags: [step({ workflow: true }), workers(workerRegistry([worker]))],
-      deps: { agent: tags.required(agentRuntime) },
-      factory: (ctx, { agent }) => agent.delegate<string, { text: string }>("unguarded-worker", ctx.input),
-    })
-    const scope = createScope({ extensions: [workflowExtension({ log: new InMemoryAgentEventLog() })] })
-    await scope.ready
-    const ctx = scope.createContext({
-      tags: [run({ taskId: "task-missing-extension", runId: "run-missing-extension" })],
-    })
-    await expect(ctx.exec({ flow: root, input: { text: "unsafe" } }))
-      .rejects.toThrow('Tag "agent.runtime" not found')
-    await ctx.close({ ok: false, error: new Error("expected") })
-    await scope.dispose()
-  })
-
-  it("rejects remote steps when no remote runner is configured", async () => {
-    const log = new InMemoryAgentEventLog()
-    const scope = createScope({ extensions: [workflowExtension({ log }), agentExtension()] })
-    await scope.ready
-    const worker = flow({
-      name: "missing-remote-runner",
-      tags: [step({ remote: true })],
-      factory: () => "local",
-    })
-
-    const ctx = scope.createContext({ tags: [run({ taskId: "task-missing-remote", runId: "run-missing-remote" })] })
-    await expect(ctx.exec({ flow: worker })).rejects.toThrow("Remote step requires remoteRunner")
-    await ctx.close({ ok: false, error: new Error("expected") })
-  })
-
-  it("routes remote work before resolving deps when runner handles it", async () => {
-    const gate = tag<string>({ label: "agent.remote.gate" })
-    const { extensions } = agent({
-      remoteRunner: {
-        run: async (event) => {
-          expect(event.targetName).toBe("remote-before-deps")
-          return "remote-only"
-        },
-      },
-    })
-    const scope = createScope({ extensions })
-    await scope.ready
-    const worker = flow({
-      name: "remote-before-deps",
-      deps: { gate: tags.required(gate) },
-      factory: (_ctx, { gate }) => gate,
-    })
-
-    const ctx = scope.createContext({ tags: [run({ taskId: "task-remote-deps", runId: "run-remote-deps" })] })
-    expect(await ctx.exec({ flow: worker, tags: [step({ remote: true })] })).toBe("remote-only")
-    await ctx.close()
-  })
-
-  it("patches JSON materials with revision conflicts", async () => {
-    const scope = createScope()
-    const ctx = scope.createContext({ tags: [run({ taskId: "task-d", runId: "run-d" })] })
-    const prStatus = material("pr-status", {
-      kind: "json",
-      initialState: { prs: {} as Record<string, { status: string }> },
-    })
-
-    const next = await patchMaterial(ctx, prStatus, [
-      { op: "add", path: "/prs/12", value: { status: "ok" } },
-    ])
-    expect(next).toEqual({
-      name: "pr-status",
-      kind: "json",
-      revision: 1,
-      state: { prs: { "12": { status: "ok" } } },
-    })
-    await expect(
-      patchMaterial(ctx, prStatus, [
-        { op: "replace", path: "/prs/12/status", value: "stale" },
-      ], { expectedRevision: 0 })
-    ).rejects.toThrow("Material revision conflict")
-    await ctx.close()
-  })
-
-  it("derives material state from primary material", async () => {
-    const source = material("count", {
-      kind: "json",
-      initialState: { value: 2 },
-    })
-    const doubled = derivedMaterial("double", source, (state) => state.value * 2, { kind: "json" })
-    const scope = createScope()
-    expect(await scope.resolve(doubled)).toEqual({
-      name: "double",
-      kind: "json",
-      revision: 0,
-      state: 4,
-    })
-  })
-
-  it("runs a real CLI worker", async () => {
-    const cli = cliWorker<{ text: string }, string>({
-      name: "printf",
-      command: "printf",
-      args: (input) => ["%s", input.text],
-      timeoutMs: 5_000,
-    })
-    const scope = createScope()
-    const ctx = scope.createContext()
-    expect(await ctx.exec({ flow: cli, input: { text: "agent-sdk-cli-ok" } })).toBe("agent-sdk-cli-ok")
-    await ctx.close()
-  })
-
-  it("marks CLI-backed LLM helpers as LLM workers", () => {
-    expect(step.find(cliWorker({ name: "x", command: "printf" })).kind).toBe("cli")
-    expect(step.find(claudeCliWorker()).kind).toBe("llm")
-    expect(step.find(codexCliWorker()).kind).toBe("llm")
-  })
-
-  it("passes LLM helper prompts after one argv terminator", async () => {
-    const scope = createScope()
-    const ctx = scope.createContext()
-
-    expect(await ctx.exec({
-      flow: claudeCliWorker({ command: "echo", extraArgs: ["--", "--verbose"] }),
-      input: { prompt: "--help" },
-    })).toBe("-p --verbose -- --help")
-    expect(await ctx.exec({
-      flow: codexCliWorker({ command: "echo", extraArgs: ["--", "--verbose"] }),
-      input: { prompt: "--help" },
-    })).toBe("exec -s read-only --verbose -- --help")
-
-    await ctx.close()
-  })
-
-  it("reports CLI failures with captured stderr", async () => {
-    const cli = cliWorker({
-      name: "sh-fail",
-      command: "sh",
-      args: ["-c", "echo bad >&2; exit 7"],
-    })
-    const scope = createScope()
-    const ctx = scope.createContext()
-    await expect(ctx.exec({ flow: cli, input: { prompt: "" } })).rejects.toMatchObject({
-      name: "CliWorkerError",
-      result: { exitCode: 7, stderr: "bad\n" },
-    } satisfies Partial<CliWorkerError>)
-    await ctx.close({ ok: false, error: new Error("expected") })
-  })
-})
