+++ packages/agent-sdk-test/tsconfig.test.json
@@ -4,5 +4,5 @@
     "rootDir": ".",
     "noEmit": true
   },
-  "include": ["src/**/*", "tests/**/*", "vitest.config.ts"]
+  "include": ["src/**/*", "examples/**/*", "tests/**/*", "vitest.config.ts"]
 }
